cleanup and moving packet table
