+++ src/gui/mod.rs
@@ -1,4 +1,3 @@
 pub mod app;
-pub mod packet_table;
 pub mod pane_tree;
 pub mod panes;
