+++ src/gui/panes/mod.rs
@@ -1,5 +1,4 @@
-use crate::gui::packet_table::PacketTable;
-
 pub mod drill_down;
 pub mod graph;
+pub mod packet_table;
 pub mod payload;
